+++ src/main/java/de/gematik/demis/igs/service/utils/RequestHeaderProvider.java
@@ -1,0 +1,59 @@
+package de.gematik.demis.igs.service.utils;
+
+/*-
+ * #%L
+ * Integrierte-Genomische-Surveillance-Service
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
+ * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
+ * European Commission – subsequent versions of the EUPL (the "Licence").
+ * You may not use this work except in compliance with the Licence.
+ *
+ * You find a copy of the Licence in the "Licence" file or at
+ * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the Licence is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either expressed or implied.
+ * In case of changes by gematik find details in the "Readme" file.
+ *
+ * See the Licence for the specific language governing permissions and limitations under the Licence.
+ *
+ * *******
+ *
+ * For additional notes and disclaimer from gematik and in case of changes by gematik find details in the "Readme" file.
+ * #L%
+ */
+
+import static de.gematik.demis.igs.service.service.validation.ValidationServiceClient.HEADER_FHIR_API_VERSION;
+import static de.gematik.demis.igs.service.service.validation.ValidationServiceClient.HEADER_FHIR_PROFILE;
+import static java.util.Optional.ofNullable;
+
+import jakarta.servlet.http.HttpServletRequest;
+import java.util.List;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.web.context.annotation.RequestScope;
+
+@Service
+@RequiredArgsConstructor
+@RequestScope
+public class RequestHeaderProvider {
+
+  private final HttpServletRequest httpServletRequest;
+
+  public @Nullable List<String> receiveApiVersionsFromRequest() {
+    return headersFromRequestByName(HEADER_FHIR_API_VERSION);
+  }
+
+  public @Nullable List<String> receiveFhirProfileFromRequest() {
+    return headersFromRequestByName(HEADER_FHIR_PROFILE);
+  }
+
+  private @Nullable List<String> headersFromRequestByName(@Nonnull String headerName) {
+    return ofNullable(httpServletRequest.getHeader(headerName)).map(List::of).orElse(null);
+  }
+}